third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-getNodeForLocation-skip-shadow.html

<html>
<head>
<script type="text/javascript" src="../../http/tests/inspector-protocol/resources/inspector-protocol-test.js"></script>
<script>

function test()
{
    var nodeInfo = {};
    InspectorTest.eventHandler["DOM.setChildNodes"] = setChildNodes;
    InspectorTest.sendCommand("DOM.enable", {});
    InspectorTest.sendCommand("DOM.getNodeForLocation", { "x": 10, "y": 10, "includeUserAgentShadowDOM": false }, onGetNodeForLocation);

    function onGetNodeForLocation(message)
    {
        if (message.error) {
            InspectorTest.log(message.error.message);
            InspectorTest.completeTest();
            return;
        }
        var nodeId = message.result.nodeId;
        InspectorTest.logObject(nodeInfo[nodeId], "Node: ", ["backendNodeId", "parentId", "nodeId"]);
        InspectorTest.completeTest();
    }

    function setChildNodes(message)
    {
        var nodes = message.params.nodes;
        for (var i = 0; i < nodes.length; ++i) {
            nodeInfo[nodes[i].nodeId] = nodes[i];
            delete nodes[i].nodeId;
        }
    }
}

</script>
</head>
<body onload="runTest()">
<form action="#">
    <input type="text" style="position:absolute;top:0;left:0;width:100;height:100" />
</form>
</body>
</html>
