third_party/WebKit/LayoutTests/inspector-protocol/dom/dom-getNodeForLocation-skip-shadow-expected.txt


Node: {
    attributes : [
        [0] : type
        [1] : text
        [2] : style
        [3] : position:absolute;top:0;left:0;width:100;height:100
    ]
    backendNodeId : <backendNodeId>
    childNodeCount : 0
    children : [
    ]
    localName : input
    nodeId : <nodeId>
    nodeName : INPUT
    nodeType : 1
    nodeValue : 
    parentId : <parentId>
    shadowRoots : [
        [0] : {
            backendNodeId : <backendNodeId>
            childNodeCount : 1
            localName : 
            nodeId : <nodeId>
            nodeName : #document-fragment
            nodeType : 11
            nodeValue : 
            shadowRootType : user-agent
        }
    ]
}
